Add PTA plots to RSLC QA output PDF report (#130)
* Add PTA plots to output PDF report

Add functionality to extract the results of the Point Target Analysis
(PTA) CalTool from the STATS.h5 file, generate plots of the azimuth &
range impulse responses for each corner reflector for each frequency &
polarization, and append the plots to the PDF report. This is a no-op if
the RSLC product did not contain any corner reflectors.

* Don't close and re-open PdfPages

Previously, only the "qa_reports" step of the RSLC QA processor wrote to
the output PDF report. With the addition of PTA plots to the PDF, we now
have two distinct steps that need to write to the same output PDF. If
both are workflows are enabled, this can cause an issue, since closing
and re-opening a `PdfPages` object causes the file to be overwritten,
discarding the previous contents.

The somewhat kludgy solution implemented here is to unconditionally
create the `PdfPages` object and keep it open during both steps. When
the PDF file is finally closed, it will be automatically deleted if
nothing was written to it. This approach avoids adding a bunch of
complicated conditional logic to the QA processor, while still ensuring
that the PDF is always safely closed. It has the downside of a creating
a messy diff (mostly whitespace changes), however.

* Cleanup style

* Use globals instead of hard-coding paths & band choices

* Simplify logic getting freq/pol child groups

* Fix capitalization of "HDF5 Group"

* Add sanity checks of STATS.h5 file structure

* Clean up long lines

* Improve docstring and comment clarity

* Capitalize the first letter in error messages

* Improve error message clarity

* Update PTA plots styling

* Use a thick solid line for power and a thin solid line for phase
* Axis/tick labels should be black to improve readability, regardless of
  the colors used in the plot
* Constrain the (power) y-axis lower limit to 50dB below the peak so
  that the nulls don't stretch the axis limits too much

* Remove unnecessary import

* Close figures after saving them to PDF

* Add PSLR/ISLR info to each plot

Only include the legend on the left sub-plot instead of both sub-plots

* Make a comment a bit more explicit

* Make `SEABORN_COLORBLIND` a public symbol

Make `seaborn_colorblind` allcaps and make it public in `nisarqa` to
reduce spaghetti code.

* Fix typo

* Clean up fstrings