Add package data manifest (#179)
This is a hack to get things working with isce3 Docker builds, which for
unknown reasons don't seem to be installing the XML files as part of the
contents of the nisarqa package.

The setuptools docs indicate that the XML files should be installed by
default if they are tracked by Git:
https://setuptools.pypa.io/en/latest/userguide/datafiles.html#include-package-data

This seems to work fine outside of Docker but mysteriously doesn't work
in isce3 Docker builds. The only solution that I was able to make work
inside the Docker image was to add a `MANIFEST.in` file that includes
the XML files.

We should revisit this if/when we begin using setuptools-scm to automate
version numbering in case that magically resolves the issue.