Integrate GSLC PTA tool (#237)
* Integrate GSLC PTA tool

Two new groups are added to the GSLC QA runconfig:

1. `dynamic_ancillary_file_group` contains an optional
   `corner_reflector_file` (required in order to run PTA) and
   `dem_file`.
2. `point_target_analyzer` contains the configuration parameters for
   the PTA tool (a subset of the RSLC PTA parameters)

A few new wrapper functions are added to run the GSLC PTA tool via isce3
for each frequency & co-polarization in the input GSLC product, and
write the results to the output QA STATS.h5 file.

The organization of the GSLC PTA results necessitates some
re-structuring of the layout of the PTA group in the STATS.h5 file.

Previously there were subgroups `azimuthIRF` and `rangeIRF` that
included metrics related to the impulse response characteristics (e.g.
ISLR, PSLR, resolution) as well as the position / geometric accuracy of
the focused peak.

In the GSLC PTA results, the impulse response characteristics still
correspond to azimuth & range cuts, but the position / geometric
accuracy metrics correspond to the image axes X & Y (not azimuth &
range). So we need to reorganize the layout to break up the contents of
these groups.

The approach in this commit is to add 4 subgroups for GSLC PTA: `x`,
`y`, `azimuth`, and `range`. The `x` and `y` groups contain metrics
related to the geolocation of the target, and the `azimuth` and `range`
groups contain info about the impulse response function.

In the RSLC PTA STATS.h5 group, there are still just 2 subgroups named
`azimuth` and `range` (dropping the `IRF` suffix) that include all of
the various metrics.

* Add GSLC PTA plots to QA report

* Add impulse response cut plots for each corner reflector
* Add scatterplot of corner reflector offsets